+++ lib/puree/event.rb
@@ -1,0 +1,88 @@
+module Puree
+
+  # Event resource
+  #
+  class Event < Resource
+
+    def initialize
+      super(:event)
+    end
+
+    # City
+    #
+    # @return [String]
+    def city
+      path = '//city'
+      xpath_query(path).text.strip
+    end
+
+    # Country
+    #
+    # @return [String]
+    def country
+      path = '//country/term/localizedString'
+      xpath_query(path).text.strip
+    end
+
+    # Date
+    #
+    # @return [Hash]
+    def date
+      data = {}
+      path = '//dateRange'
+      range = xpath_query path
+      data['start'] = range.xpath('startDate').text.strip
+      data['end'] = range.xpath('startDate').text.strip
+      data
+    end
+
+    # Description
+    #
+    # @return [String]
+    def description
+      path = '//content/description'
+      xpath_query(path).text.strip
+    end
+
+    # Location
+    #
+    # @return [String]
+    def location
+      path = '//location'
+      xpath_query(path).text.strip
+    end
+
+    # Title
+    #
+    # @return [String]
+    def title
+      path = '//title/localizedString'
+      xpath_query(path).text.strip
+    end
+
+    # Type
+    #
+    # @return [String]
+    def type
+      path = '//content/typeClassification/term/localizedString'
+      xpath_query(path).text.strip
+    end
+
+    # All metadata
+    #
+    # @return [Hash]
+    def metadata
+      o = {}
+      o['city'] = city
+      o['country'] = country
+      o['date'] = date
+      o['description'] = description
+      o['location'] = location
+      o['title'] = title
+      o['type'] = type
+      o
+    end
+
+  end
+
+end